+++ parsers/LitCoin/src/bagel/bagel.py
@@ -4,7 +4,7 @@ from requests.adapters import HTTPAdapter, Retry
 
 from parsers.LitCoin.src.NER.nameres import NameResNEREngine
 from parsers.LitCoin.src.NER.sapbert import SAPBERTNEREngine
-from parsers.LitCoin.src.bagel_gpt import ask_classes_and_descriptions
+from parsers.LitCoin.src.bagel.bagel_gpt import ask_classes_and_descriptions
 from Common.normalization import NODE_NORMALIZATION_URL
 
 # output of parse_gpt looks like {"entity": triple["object"], "qualifier": triple["object_qualifier"]}
